--- ckanext-basedosdados/ckanext/basedosdados/validator/available_options/admin1.py
@@ -1,33 +1,0 @@
-from .attr_enum import AttrEnum
-
-class Admin1Enum(AttrEnum):
-    
-    # por enquanto só para Brasil
-
-    id_uf_12 = {"label": "AC"}
-    id_uf_27 = {"label": "AL"}
-    id_uf_13 = {"label": "AM"}
-    id_uf_16 = {"label": "AP"}
-    id_uf_29 = {"label": "BA"}
-    id_uf_23 = {"label": "CE"}
-    id_uf_53 = {"label": "DF"}
-    id_uf_32 = {"label": "ES"}
-    id_uf_52 = {"label": "GO"}
-    id_uf_21 = {"label": "MA"}
-    id_uf_31 = {"label": "MG"}
-    id_uf_50 = {"label": "MS"}
-    id_uf_51 = {"label": "MT"}
-    id_uf_15 = {"label": "PA"}
-    id_uf_25 = {"label": "PB"}
-    id_uf_26 = {"label": "PE"}
-    id_uf_22 = {"label": "PI"}
-    id_uf_41 = {"label": "PR"}
-    id_uf_33 = {"label": "RJ"}
-    id_uf_24 = {"label": "RN"}
-    id_uf_11 = {"label": "RO"}
-    id_uf_14 = {"label": "RR"}
-    id_uf_43 = {"label": "RS"}
-    id_uf_42 = {"label": "SC"}
-    id_uf_28 = {"label": "SE"}
-    id_uf_35 = {"label": "SP"}
-    id_uf_17 = {"label": "TO"}
