--- ckanext-basedosdados/ckanext/basedosdados/validator/available_options/admin2.py
@@ -1,34 +1,0 @@
-
-from .attr_enum import AttrEnum
-
-class Admin2Enum(AttrEnum):
-    
-    # por enquanto só para capitais de UFs no Brasil
-    
-    id_municipio_1100205 = {"label": "Porto Velho"}
-    id_municipio_1200401 = {"label": "Rio Branco"}
-    id_municipio_1302603 = {"label": "Manaus"}
-    id_municipio_1400100 = {"label": "Boa Vista"}
-    id_municipio_1501402 = {"label": "Belém"}
-    id_municipio_1600303 = {"label": "Macapá"}
-    id_municipio_1721000 = {"label": "Palmas"}
-    id_municipio_2111300 = {"label": "São Luís"}
-    id_municipio_2211001 = {"label": "Teresina"}
-    id_municipio_2304400 = {"label": "Fortaleza"}
-    id_municipio_2408102 = {"label": "Natal"}
-    id_municipio_2507507 = {"label": "João Pessoa"}
-    id_municipio_2611606 = {"label": "Recife"}
-    id_municipio_2704302 = {"label": "Maceió"}
-    id_municipio_2800308 = {"label": "Aracaju"}
-    id_municipio_2927408 = {"label": "Salvador"}
-    id_municipio_3106200 = {"label": "Belo Horizonte"}
-    id_municipio_3205309 = {"label": "Vitória"}
-    id_municipio_3304557 = {"label": "Rio de Janeiro"}
-    id_municipio_3550308 = {"label": "São Paulo"}
-    id_municipio_4106902 = {"label": "Curitiba"}
-    id_municipio_4205407 = {"label": "Florianópolis"}
-    id_municipio_4314902 = {"label": "Porto Alegre"}
-    id_municipio_5002704 = {"label": "Campo Grande"}
-    id_municipio_5103403 = {"label": "Cuiabá"}
-    id_municipio_5208707 = {"label": "Goiânia"}
-    id_municipio_5300108 = {"label": "Brasília"}
