--- ckanext-basedosdados/ckanext/basedosdados/validator/available_options/continent.py
@@ -1,14 +1,0 @@
-from .attr_enum import AttrEnum
-
-class ContinentEnum(AttrEnum):
-    
-    all             = {"label": "Todos"}
-    africa          = {"label": "África"}
-    north_america   = {"label": "América do Norte"}
-    central_america = {"label": "América Central"}
-    south_america   = {"label": "América do Sul"}
-    asia            = {"label": "Ásia"}
-    europe          = {"label": "Europa"}
-    oceania         = {"label": "Oceania"}
-
-
