+++ ckanext-basedosdados/ckanext/basedosdados/validator/resources/bdm/column/__init__.py
@@ -23,8 +23,7 @@ class BdmColumns(BaseModel):
     covered_by_dictionary: Optional[YesNoEnum]           = COVERED_BY_DICTIONARY_FIELD
     directory_column     : Optional[DirectoryColumn]     = DIRECTORY_COLUMN_FIELD
     measurement_unit     : Optional[MeasurementUnitEnum] = MEASUREMENT_UNIT_FIELD
-    #original_names       : Optional[List[Str]]           = ORIGINAL_NAMES_FIELD
-        #TODO: make it a complex field of dictionaries mapping year to name. Ex: {2015: "name_1", 2016: "name_2"}
+    #original_names       : Optional[List[Str]]           = ORIGINAL_NAMES_FIELD #TODO: make it a complex field of dictionaries mapping year to name. Ex: {2015: "name_1", 2016: "name_2"}
 
     is_in_staging        : Optional[bool]                = IS_IN_STAGING_FIELD
     is_partition         : Optional[bool]                = IS_PARTITION_FIELD
