+++ ckanext-basedosdados/ckanext/basedosdados/validator/resources/bdm/table/fields_definitions.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 from datetime import datetime
 from typing import Optional, Set
+
 from pydantic import (
     StrictStr as Str,
     Field,
@@ -23,7 +24,6 @@ class SpatialCoverage(BaseModel):
     # 1. ler os dataframes de diretorios para estrurar árvore de dicts e metadados
         # incluir IDs de entidades e nomes
     # 2. transformar isso num dict para front-end
-
     continent : Optional[Set[ContinentEnum]] = Field(user_input_hint=["Continente"])
     country   : Optional[Set[CountryEnum]]   = Field(user_input_hint=["País"])
     admin1    : Optional[Set[Admin1Enum]]    = Field(user_input_hint=["UF/Estado"])
